Refactor routing into a separate file
